all_requirements.txt
alabaster==0.7.12
appnope==0.1.0
argon2-cffi==19.1.0
arrow==0.13.1
atomicwrites==1.3.0
attrs==19.1.0
awsebcli==3.15.0
Babel==2.6.0
backcall==0.1.0
beautifulsoup4==4.6.0
binaryornot==0.4.4
blessed==1.15.0
botocore==1.12.140
cached-property==1.5.1
cement==2.8.2
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
Click==7.0
Collectfast==0.6.2
colorama==0.3.9
cookiecutter==1.6.0
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.3
crispy-forms-foundation==0.7.0
decorator==4.4.0
defusedxml==0.6.0
Django==2.2.1
django-admin-tools==0.8.1
django-allauth==0.39.1
django-any-imagefield==1.0
django-any-urlfield==2.6.1
django-anymail==6.0
django-appconf==1.0.3
django-categories-i18n==1.1
django-ckeditor==5.7.0
django-compressor==2.2
django-contrib-comments==1.9.1
django-coverage-plugin==1.6.0
django-crispy-forms==1.7.2
django-debug-toolbar==1.11
django-disqus==0.5
django-environ==0.4.5
django-extensions==2.1.6
django-filer==1.5.0
django-fluent-blogs==2.0.3
django-fluent-comments==2.1
django-fluent-contents==2.0.5
django-fluent-dashboard==1.0
django-fluent-pages==2.0.5
django-fluent-utils==2.0.1
django-foundation-formtags==0.1.0
django-guardian==1.5.1
django-js-asset==1.2.2
django-libsass==0.7
django-model-utils==3.1.2
django-modelcluster==4.4
django-mptt==0.10.0
django-parler==1.9.2
django-phonenumber-field==2.3.1
django-polymorphic==2.0.3
django-polymorphic-tree==1.5
django-redis==4.10.0
django-slug-preview==1.0.4
django-staff-toolbar==1.2
django-storages==1.7.1
django-tag-parser==3.1
django-taggit==0.24.0
django-taggit-autosuggest==0.3.7
django-template-analyzer==1.6.1
django-threadedcomments==1.2
django-treebeard==4.3
django-wysiwyg==0.8.0
djangorestframework==3.9.3
docker==3.7.2
docker-compose==1.23.2
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
docutils==0.14
draftjs-exporter==2.1.5
easy-thumbnails==2.6
entrypoints==0.3
factory-boy==2.11.1
Faker==1.0.5
flake8==3.7.7
fluentcms-campaign==0.9.4
fluentcms-contactform==2.0
fluentcms-cookielaw==2.0
fluentcms-countdown==1.2
fluentcms-emailtemplates==2.0
fluentcms-file==1.0.2
fluentcms-filer==1.1.1
fluentcms-teaser==1.1.2
future==0.16.0
gunicorn==19.9.0
html2text==2018.1.9
html5lib==1.0.1
idna==2.7
imagesize==1.1.0
ipdb==0.12
ipython==7.5.0
ipython-genutils==0.2.0
itypes==1.1.0
jedi==0.13.3
Jinja2==2.10.1
jinja2-time==0.2.0
jmespath==0.9.4
jsonschema==2.6.0
libsass==0.18.0
Markdown==3.1
MarkupSafe==1.1.1
mccabe==0.6.1
micawber==0.4.0
mock==2.0.0
more-itertools==7.0.0
mypy==0.701
mypy-extensions==0.4.1
oauthlib==3.0.1
packaging==19.0
parso==0.4.0
pathspec==0.5.9
pbr==5.2.0
pexpect==4.7.0
pickleshare==0.7.5
Pillow==5.4.1
pipenv==2018.11.26
pluggy==0.9.0
poyo==0.4.2
prompt-toolkit==2.0.9
psycopg2==2.8.2
psycopg2-binary==2.8.2
ptyprocess==0.6.0
py==1.8.0
pycodestyle==2.5.0
pycparser==2.19
pyflakes==2.1.1
Pygments==2.3.1
pyparsing==2.4.0
pytest==4.4.1
pytest-django==3.4.8
pytest-sugar==0.9.2
python-akismet==0.4.1
python-dateutil==2.8.0
python-decouple==3.1
python-slugify==3.0.2
python3-openid==3.1.0
pytz==2019.1
PyYAML==3.13
rcssmin==1.0.6
redis==3.2.1
requests==2.20.1
requests-oauthlib==1.2.0
rjsmin==1.0.12
semantic-version==2.5.0
six==1.11.0
snowballstemmer==1.2.1
Sphinx==2.0.1
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
sqlparse==0.3.0
swapper==1.1.0
termcolor==1.1.0
text-unidecode==1.2
textile==3.0.3
texttable==0.9.1
traitlets==4.3.2
twitter-text==3.0
typed-ast==1.3.5
Unidecode==1.0.23
uritemplate==3.0.0
urllib3==1.24.2
virtualenv==16.5.0
virtualenv-clone==0.5.3
wagtail==2.5
wcwidth==0.1.7
webencodings==0.5.1
websocket-client==0.56.0
Werkzeug==0.15.2
whichcraft==0.5.2
Willow==1.1
WsgiUnproxy==1.0
